@@ -1,6 +1,4 @@
-import 'package:test/test.dart';
-
-import 'package:flutter_page_indicator/flutter_page_indicator.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   test('adds one to input values', () {});
